+++ ft_ls.h
@@ -6,7 +6,7 @@
 /*   By: dslogrov <dslogrove@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/21 16:24:11 by dslogrov          #+#    #+#             */
-/*   Updated: 2018/07/18 16:54:45 by dslogrov         ###   ########.fr       */
+/*   Updated: 2018/07/24 11:28:20 by dslogrov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,12 +80,12 @@ typedef struct		s_file_info
 	struct stat		stat;
 }					t_file_info;
 
-int					compare_default(const t_file_info *a, const t_file_info *b);
-int					compare_lc(const t_file_info *a, const t_file_info *b);
-int					compare_lu(const t_file_info *a, const t_file_info *b);
-int					compare_uu(const t_file_info *a, const t_file_info *b);
-int					compare_us(const t_file_info *a, const t_file_info *b);
-int					compare_lt(const t_file_info *a, const t_file_info *b);
+long				compare_default(const t_file_info *a, const t_file_info *b);
+long				compare_lc(const t_file_info *a, const t_file_info *b);
+long				compare_lu(const t_file_info *a, const t_file_info *b);
+long				compare_uu(const t_file_info *a, const t_file_info *b);
+long				compare_us(const t_file_info *a, const t_file_info *b);
+long				compare_lt(const t_file_info *a, const t_file_info *b);
 
 void				ls_print_normal(const t_list *list, t_flag flags);
 void				ls_print_ll(const t_list *list, t_flag flags);
